chore: refactor folder structure
