+++ src/pages/home/_testimonial-carousel.tsx
@@ -2,10 +2,10 @@ import React, { Children, useState, useEffect, useRef, ReactNode } from 'react'
 import type { Dispatch, SetStateAction } from 'react'
 import styled from 'styled-components'
 import { Flex } from 'components/containers'
-import ArrowLeft from 'images/svg/testimonials/arrow-left.svg'
-import ArrowRight from 'images/svg/testimonials/arrow-right.svg'
-import ArrowLeftFade from 'images/svg/testimonials/arrow-left-fade.svg'
-import ArrowRightFade from 'images/svg/testimonials/arrow-right-fade.svg'
+import ArrowLeft from 'images/svg/testimonials/testimonial_arrow_left.svg'
+import ArrowRight from 'images/svg/testimonials/testimonial_arrow_right.svg'
+import ArrowLeftFade from 'images/svg/testimonials/testimonial_arrow_left_fade.svg'
+import ArrowRightFade from 'images/svg/testimonials/testimonial_arrow_right_fade.svg'
 import device from 'themes/device'
 import { useIsRtl } from 'components/hooks/use-isrtl'
