+++ src/pages/terms-and-conditions/_business-grid.tsx
@@ -5,14 +5,14 @@ import { localize, Localize } from 'components/localization'
 import { getLanguage } from 'common/utility'
 import { TString } from 'types/generics'
 // Icons
-import General from 'images/svg/terms/business-general-tc.svg'
-import Partners from 'images/svg/terms/business-partners-tc.svg'
-import PA from 'images/svg/terms/payment-agents.svg'
-import API from 'images/svg/terms/business-api-tc.svg'
-import APIROW from 'images/svg/terms/api-users-row.svg'
-import BugBounty from 'images/svg/terms/bug-bounty.svg'
-import BugBountyRow from 'images/svg/terms/bug-bounty-row.svg'
-import IG from 'images/svg/terms/important-guidelines.svg'
+import General from 'images/svg/terms/term_business_general_tc.svg'
+import Partners from 'images/svg/terms/term_business_partners_tc.svg'
+import PA from 'images/svg/terms/term_payment_agents.svg'
+import API from 'images/svg/terms/term_business_api_tc.svg'
+import APIROW from 'images/svg/terms/term_api_users_row.svg'
+import BugBounty from 'images/svg/terms/term_bug_bounty.svg'
+import BugBountyRow from 'images/svg/terms/term_bug_bounty_row.svg'
+import IG from 'images/svg/terms/term_important_guidelines.svg'
 import PDF from 'images/svg/regulatory/regulatory_pdf_icon_black.svg'
 import useRegion from 'components/hooks/use-region'
