+++ src/pages/terms-and-conditions/_clients-grid.tsx
@@ -3,13 +3,13 @@ import { StyledGrid, StyledContainer, IconWrapper, GridCol, Cta } from './_terms
 import { Header, Text } from 'components/elements'
 import { Localize, localize } from 'components/localization'
 // Icons
-import General from 'images/svg/terms/general-tc.svg'
-import Trading from 'images/svg/terms/trading-tc.svg'
-import Funds from 'images/svg/terms/funds-transfers-tc.svg'
-import Security from 'images/svg/terms/security-privacy-tc.svg'
-import Risk from 'images/svg/terms/risk-tc.svg'
+import General from 'images/svg/terms/term_general_tc.svg'
+import Trading from 'images/svg/terms/term_trading_tc.svg'
+import Funds from 'images/svg/terms/term_funds_transfers_tc.svg'
+import Security from 'images/svg/terms/term_security_privacy_tc.svg'
+import Risk from 'images/svg/terms/term_risk_tc.svg'
 import PDF from 'images/svg/regulatory/regulatory_pdf_icon_black.svg'
-import BFX from 'images/svg/terms/bfx-tc.svg'
+import BFX from 'images/svg/terms/term_bfx_tc.svg'
 import useRegion from 'components/hooks/use-region'
 import { TString } from 'types/generics'
